refactor(raf most stream): moved raf most stream to correct folder, tried different implementation
